demo_requirements.txt
# Demo Pipeline Requirements
# Core dependencies for the integrated emotion recognition system

# Computer Vision and Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Deep Learning and ML
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Progress and Utilities
tqdm>=4.62.0

# Optional: For better performance
# torchaudio  # If using audio features
# scikit-learn  # For additional ML utilities
